_scripts/get_release_info/pyproject.toml
[project]
name = "get-release-info"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]
